% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{AStrain}
\alias{AStrain}
\title{Training models for posterior selection of algorithms}
\usage{
AStrain(data_object, ...)
}
\arguments{
\item{data_object}{an object.}

\item{...}{other parameters.}
}
\value{
A list, result of the respective AStrain method.
}
\description{
For each algorithm (column) in the data, a model is trained to later predict the output (KPI) for that algorithm (using function \code{ASpredict()}).
}
