% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{boxplots}
\alias{boxplots}
\title{Boxplots}
\usage{
boxplots(data_object, ...)
}
\arguments{
\item{data_object}{an object.}

\item{...}{other parameters.}
}
\value{
A \code{ggplot} object, result of the respective boxplots method.
}
\description{
Generates boxplots for an object.
}
