% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{figure_comparison}
\alias{figure_comparison}
\title{Figure comparison}
\usage{
figure_comparison(data_object, ...)
}
\arguments{
\item{data_object}{an object}

\item{...}{other parameters}
}
\value{
A \code{ggplot} object, result of the respective figure_comparison method.
}
\description{
Generates figure comparison plot for an object.
}
