% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ranking}
\alias{ranking}
\title{Ranking}
\usage{
ranking(data_object, ...)
}
\arguments{
\item{data_object}{an object}

\item{...}{other parameters}
}
\value{
A \code{ggplot} object, result of the respective ranking method.
}
\description{
Generates ranking plot for an object.
}
