% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpMVformat}
\alias{SpMVformat}
\title{Automatic selection of the most suitable storage format for sparse matrices on GPUs}
\format{
A list with x (features) and y (KPIs) data.frames.
}
\source{
Pichel, J. C., & Pateiro-López, B. (2018). A new approach for sparse matrix classification based on deep learning techniques. In 2018 IEEE International Conference on Cluster Computing (CLUSTER) (pp. 46–54).
}
\usage{
SpMVformat
}
\description{
Data from Pichel and Pateiro-López (2018), which contains information on 8111 sparse matrices. Each matrix is described by a set of nine structural features, and the performance of the single-precision SpMV kernel was measured under three storage formats: compressed row storage (CSR), ELLPACK (ELL), and hybrid (HYB). For each matrix and format, performance is expressed as the average GFLOPS (billions of floating-point operations per second), over 1000 SpMV operations.
}
\keyword{datasets}
