% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{ASpredict}
\alias{ASpredict}
\title{Predicting the KPI value for the algorithms}
\usage{
ASpredict(training_object, ...)
}
\arguments{
\item{training_object}{list of class \code{as_train}.}

\item{...}{other parameters.}
}
\value{
A data frame, result of the respective ASpredict method.
}
\description{
For each algorithm, the output (KPI) is predicted using the models trained with \code{AStrain()}.
}
