% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{KPI_table}
\alias{KPI_table}
\title{Internal generic for KPI_table}
\usage{
KPI_table(data_object, ...)
}
\arguments{
\item{data_object}{object.}

\item{...}{other parameters.}
}
\description{
This function serves as the internal S3 generic for \code{KPI_table} methods.
It dispatches the call to the appropriate method based on the class of \code{data_object}.
Currently, only \code{as_data} is implemented. Users or developers can
extend this generic by writing new methods for other classes.
}
\details{
This generic is not intended to be used directly by package users. It exists
to enable method dispatch for different classes. Marked as \code{internal} to keep
it out of the user-facing function index.
}
\keyword{internal}
